refactor(health): extract types + formatters from health-page
WS6 client-component split, following the audit's seam map. Two pure, leaf
modules move out of the 1,155-line health-page.tsx:

- plugins/health/types.ts — the 14 wire-format interfaces (McpSessionInfo,
  DoctorData, ServerData, PluginInfo, RegistryData, PluginManifestEntry/Data,
  ErrorsByKind, UsageEntry/Kind, TopByNameRow, ByAgentRow, UsageFeedData,
  HealthSummary). Fixes the missing-types.ts convention violation (tasks/models
  plugins both have one); plugins/health/index.ts can now type its route
  payloads against the same contracts.
- plugins/health/lib/format.ts — the pure formatters (formatUptime,
  formatTokenCount, formatRuntimeCost, formatDateShort, searchStatsDocumentCount,
  extractErrorMessage, formatActivity). formatAge already comes from
  @makinbakin/sdk/utils — not duplicated.

health-page.tsx: 1,155 -> 996 lines. Pure relocation — types are erased at
build, formatters are pure, and the HealthPage export + client.tsx slot
registration are untouched (types <- lib/format <- component is a DAG, no cycles).

Deferred (the involved React-state work): the section-component split
(usage/plugins/search/agent-usage), the per-section-fetch redesign, and
promoting the inline searchHealth state type to a named SearchHealthData.

Verified: typecheck + lint clean; health suite 98/0 (incl. health-page.test.tsx);
full suite 5123/0; binary build; isolated-home boot smoke (Health activates;
GET /api/plugins/health/summary -> 200; client.js bundle -> 200).
